Refactor dummy item generation
Extract item generation into separate class to make it reusable.
